[Added] Docs Swagger for Product

+++ app/Models/Product.php
@@ -7,6 +7,37 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
+/**
+ * Class Product
+ *
+ * @OA\Schema(
+ *     @OA\Xml(name="Product"),
+ *     description="Product Model",
+ *     type="object",
+ *     title="Product Model",
+ *     @OA\Property(property="id", type="integer", format="int64"),
+ *     @OA\Property(property="product_name", type="string"),
+ *     @OA\Property(property="product_desc", type="string"),
+ *     @OA\Property(property="product_release_date", type="string", format="date"),
+ *     @OA\Property(property="product_acquisition_date", type="string", format="date"),
+ *     @OA\Property(property="product_qty", type="integer", format="int64"),
+ *     @OA\Property(property="product_acquisition_cost", type="number", format="decimal"),
+ *     @OA\Property(property="created_at", type="string", format="date"),
+ *     @OA\Property(property="updated_at", type="string", format="date"),
+ *     @OA\Property(property="deleted_at", type="string", format="date"),
+ * )
+ *
+ * @property int $id
+ * @property string $product_name
+ * @property string $product_desc
+ * @property string $product_release_date
+ * @property string $product_acquisition_date
+ * @property int $product_qty
+ * @property decimal $product_acquisition_cost
+ * @property string $created_at
+ * @property string $updated_at
+ * @property string $deleted_at
+ */
 class Product extends Model
 {
     use HasFactory, SoftDeletes;
@@ -31,6 +62,6 @@ class Product extends Model
     }
 
     protected $casts = [
-        'product_acquisition_cost' => 'integer',
+        'product_acquisition_cost' => 'float',
     ];
 }
